refs #18 updated change status view for tour participations
The change status view for tour participations now informs the tour guide
whether an email will be sent to the participant dependent on the presence
of an email address in the current tour participation.
The text area the allows the tour guide to enter an optional message is
only shown, when an email address is present.