refs #244 added testMode parameter
A new parameter -testMode for the sendTourChangesNotifications tourdb
shell command allows to abort the normal sending of notification emails.
If the parameter is just set (no argument), no email will be send. If it
is set to a string, the shell assumes an email address and sends the
notification to this address only.